moved AbstractNode from dag.gsoc package to dag
