+++ engine/src/main/java/org/terasology/rendering/dag/RenderGraph.java
@@ -31,7 +31,6 @@ import org.terasology.engine.SimpleUri;
 import org.terasology.naming.Name;
 import org.terasology.rendering.ShaderManager;
 import org.terasology.rendering.dag.gsoc.*;
-import org.terasology.rendering.dag.gsoc.AbstractNode;
 
 /**
  * TODO: Add javadocs
