+++ engine/src/main/java/org/terasology/rendering/dag/nodes/RefractiveReflectiveBlocksNodeProxy.java
@@ -18,7 +18,7 @@ package org.terasology.rendering.dag.nodes;
 import org.terasology.context.Context;
 import org.terasology.engine.SimpleUri;
 import org.terasology.naming.Name;
-import org.terasology.rendering.dag.gsoc.AbstractNode;
+import org.terasology.rendering.dag.AbstractNode;
 import org.terasology.rendering.nui.properties.Range;
 import java.beans.PropertyChangeEvent;
 import java.beans.PropertyChangeListener;
